Fix shallow copy for superbias and wavecorr; add tests
