+++ jwst/superbias/superbias_step.py
@@ -36,6 +36,9 @@ class SuperBiasStep(Step):
         """
         # Open the input data model
         with datamodels.open(step_input) as input_model:
+            # Work on a copy
+            result = input_model.copy()
+
             # Get the name of the superbias reference file to use
             self.bias_name = self.get_reference_file(input_model, "superbias")
             log.info("Using SUPERBIAS reference file %s", self.bias_name)
@@ -44,14 +47,11 @@ class SuperBiasStep(Step):
             if self.bias_name == "N/A":
                 log.warning("No SUPERBIAS reference file found")
                 log.warning("Superbias step will be skipped")
-                input_model.meta.cal_step.superbias = "SKIPPED"
-                return input_model
+                result.meta.cal_step.superbias = "SKIPPED"
+                return result
 
             # Open the superbias ref file data model
             bias_model = datamodels.SuperBiasModel(self.bias_name)
-
-            # Work on a copy
-            result = input_model.copy()
 
             # Do the bias subtraction
             result = bias_sub.do_correction(result, bias_model)
