+++ jwst/wavecorr/tests/test_wavecorr.py
@@ -60,6 +60,11 @@ def test_wavecorr():
     im_src.slits[0].source_type = "POINT"
     im_wave = WavecorrStep.call(im_src)
 
+    # step is marked complete; input is not modified
+    assert im_wave.meta.cal_step.wavecorr == "COMPLETE"
+    assert im_src is not im_wave
+    assert im_src.meta.cal_step.wavecorr is None
+
     # test dispersion is of the correct order
     # there's one slit only
     slit = im_src.slits[0]
@@ -120,8 +125,7 @@ def test_ideal_to_v23_fs():
     assert_allclose(id2v.inverse(v2_ref, v3_ref), (0, 0))
 
 
-def test_skipped():
-    """Test all conditions that lead to skipping wavecorr."""
+def test_skip_invalid_exptype():
     hdul = create_nirspec_fs_file(grating="G140H", filter="F100LP")
     im = datamodels.ImageModel(hdul)
 
@@ -130,10 +134,24 @@ def test_skipped():
     out = WavecorrStep.call(im)
     assert out.meta.cal_step.wavecorr == "SKIPPED"
 
+    # input is not modified
+    assert out is not im
+    assert im.meta.cal_step.wavecorr is None
+
+
+def test_skip_missing_prerequisites():
+    hdul = create_nirspec_fs_file(grating="G140H", filter="F100LP")
+    im = datamodels.ImageModel(hdul)
+
     # Test an error is raised if assign_wcs or extract_2d were not run.
     im.meta.exposure.type = "NRS_FIXEDSLIT"
     with pytest.raises(TypeError):
         WavecorrStep.call(im)
+
+
+def test_reference_file_requirements():
+    hdul = create_nirspec_fs_file(grating="G140H", filter="F100LP")
+    im = datamodels.ImageModel(hdul)
 
     outa = AssignWcsStep.call(im)
 
@@ -167,10 +185,15 @@ def test_skipped():
     outs.meta.observation.date = "2001-08-03"
     outw = WavecorrStep.call(outs)
     assert outw.meta.cal_step.wavecorr == "SKIPPED"
-    outs.meta.observation.date = "2017-08-03"
+    assert outw is not outs
+    assert outs.meta.cal_step.wavecorr is None
 
     # Run the step for real
+    outs.meta.observation.date = "2017-08-03"
     outw = WavecorrStep.call(outs)
+    assert outw.meta.cal_step.wavecorr == "COMPLETE"
+    assert outw is not outs
+    assert outs.meta.cal_step.wavecorr is None
 
     # Test if the corrected wavelengths are not monotonically increasing
 
@@ -216,6 +239,10 @@ def test_mos_slit_status():
     # since no slits were corrected
     assert im_wave.meta.cal_step.wavecorr == "SKIPPED"
 
+    # input is not modified
+    assert im_wave is not im_src
+    assert im_src.meta.cal_step.wavecorr is None
+
     # check that wavelength_corrected is False for extended mos sources
     assert im_wave.slits[0].wavelength_corrected is False
 
@@ -228,6 +255,10 @@ def test_mos_slit_status():
 
     # check that wavelength_corrected is True for mos point sources
     assert im_wave.slits[0].wavelength_corrected is True
+
+    # input is still not modified
+    assert im_wave is not im_src
+    assert im_src.meta.cal_step.wavecorr is None
 
 
 def test_wavecorr_fs():
@@ -323,6 +354,10 @@ def test_assign_wcs_skipped():
     result = WavecorrStep.call(im)
     assert result.meta.cal_step.wavecorr == "SKIPPED"
 
+    # input is not modified
+    assert result is not im
+    assert im.meta.cal_step.wavecorr is None
+
     hdul.close()
     im.close()
     result.close()
@@ -354,6 +389,10 @@ def test_invalid_slit(nrs_slit_model):
     result = WavecorrStep.call(slit)
     assert result.meta.cal_step.wavecorr == "SKIPPED"
 
+    # input is not modified
+    assert result is not slit
+    assert slit.meta.cal_step.wavecorr is None
+
 
 @pytest.mark.parametrize("source_type", ["POINT", "EXTENDED"])
 def test_slitmodel(source_type, nrs_slit_model):
@@ -365,6 +404,10 @@ def test_slitmodel(source_type, nrs_slit_model):
         assert result.meta.cal_step.wavecorr == "COMPLETE"
     else:
         assert result.meta.cal_step.wavecorr == "SKIPPED"
+
+    # input is not modified
+    assert result is not slit
+    assert slit.meta.cal_step.wavecorr is None
 
     slit.close()
     result.close()
